# FilmApi/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FilmApi",
    "version": "v1"
  },
  "paths": {
    "/Books/getBookByISBNExternal/{isbn}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "isbn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Books/GetConnectedGamesIds/{isbn}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "isbn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Books/GetConnectedFilmsIds/{isbn}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "isbn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Books/getBookByISBN/{isbn}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "isbn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Books/createBook": {
      "post": {
        "tags": [
          "Books"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateBookBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpResponseMessage"
            }
          }
        }
      }
    },
    "/Books/deleteBookByISBN/{id}": {
      "delete": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Books/updateBook/{id}": {
      "put": {
        "tags": [
          "Books"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookBeautifulResponse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/GetAllFavouriteBooks/{ChatId}": {
      "get": {
        "tags": [
          "Favourites"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ChatId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FavouriteBook"
              }
            }
          }
        }
      }
    },
    "/Favourites/createFavouriteBook": {
      "post": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteBookBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/deleteFavouriteBook": {
      "delete": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteBookBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/GetAllFavouriteGames/{ChatId}": {
      "get": {
        "tags": [
          "Favourites"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ChatId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FavouriteGame"
              }
            }
          }
        }
      }
    },
    "/Favourites/createFavouriteGame": {
      "post": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteGameBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/deleteFavouriteGame": {
      "delete": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteGameBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/GetAllFavouriteFilms/{ChatId}": {
      "get": {
        "tags": [
          "Favourites"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ChatId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FavouriteFilm"
              }
            }
          }
        }
      }
    },
    "/Favourites/createFavouriteFilm": {
      "post": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteFilmBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Favourites/deleteFavouriteFilm": {
      "delete": {
        "tags": [
          "Favourites"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FavouriteFilmBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateBookBody": {
      "type": "object",
      "properties": {
        "isbn_10": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "type": "string"
        },
        "publish_date": {
          "type": "string"
        }
      }
    },
    "Version": {
      "type": "object",
      "properties": {
        "major": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minor": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "build": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "revision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "majorRevision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minorRevision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "StringStringIEnumerableKeyValuePair": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string",
          "readOnly": true
        },
        "value": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "readOnly": true
        }
      }
    },
    "HttpContent": {
      "type": "object",
      "properties": {
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        }
      }
    },
    "HttpStatusCode": {
      "format": "int32",
      "enum": [
        100,
        101,
        102,
        103,
        200,
        201,
        202,
        203,
        204,
        205,
        206,
        207,
        208,
        226,
        300,
        301,
        302,
        303,
        304,
        305,
        306,
        307,
        308,
        400,
        401,
        402,
        403,
        404,
        405,
        406,
        407,
        408,
        409,
        410,
        411,
        412,
        413,
        414,
        415,
        416,
        417,
        421,
        422,
        423,
        424,
        426,
        428,
        429,
        431,
        451,
        500,
        501,
        502,
        503,
        504,
        505,
        506,
        507,
        508,
        510,
        511
      ],
      "type": "integer"
    },
    "HttpVersionPolicy": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "HttpMethod": {
      "type": "object",
      "properties": {
        "method": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "HttpRequestMessage": {
      "type": "object",
      "properties": {
        "version": {
          "$ref": "#/definitions/Version"
        },
        "versionPolicy": {
          "$ref": "#/definitions/HttpVersionPolicy"
        },
        "content": {
          "$ref": "#/definitions/HttpContent"
        },
        "method": {
          "$ref": "#/definitions/HttpMethod"
        },
        "requestUri": {
          "format": "uri",
          "type": "string"
        },
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "properties": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        },
        "options": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        }
      }
    },
    "HttpResponseMessage": {
      "type": "object",
      "properties": {
        "version": {
          "$ref": "#/definitions/Version"
        },
        "content": {
          "$ref": "#/definitions/HttpContent"
        },
        "statusCode": {
          "$ref": "#/definitions/HttpStatusCode"
        },
        "reasonPhrase": {
          "type": "string"
        },
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "trailingHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "requestMessage": {
          "$ref": "#/definitions/HttpRequestMessage"
        },
        "isSuccessStatusCode": {
          "type": "boolean",
          "readOnly": true
        }
      }
    },
    "BookBeautifulResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "type": "string"
        },
        "publish_date": {
          "type": "string"
        },
        "isbn_10": {
          "type": "string"
        }
      }
    },
    "FavouriteBook": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "chatId": {
          "type": "string"
        },
        "book_id": {
          "type": "string"
        }
      }
    },
    "FavouriteBookBody": {
      "type": "object",
      "properties": {
        "chatId": {
          "type": "string"
        },
        "book_id": {
          "type": "string"
        }
      }
    },
    "FavouriteGame": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "chatId": {
          "type": "string"
        },
        "game_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "FavouriteGameBody": {
      "type": "object",
      "properties": {
        "chatId": {
          "type": "string"
        },
        "game_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "FavouriteFilm": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "chatId": {
          "type": "string"
        },
        "film_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "FavouriteFilmBody": {
      "type": "object",
      "properties": {
        "chatId": {
          "type": "string"
        },
        "film_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}